feat: refactor OceanUnorderedList and OceanCalloutList
